requirements.txt
pandas
numpy
torch
scikit-learn
yfinance
matplotlib
seaborn
alpaca-trade-api
PyYAML
